+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Indeterminate state support
+
+## Changed
+
+**BREAKING** - Rename `toggle` method to `toggleAll` to check all checkboxes.
+
 ## [1.0.0] - 2020-10-15
 
 ### Added